voice-assistant/pyproject.toml
[project]
name = "nlp-audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edge-tts>=7.2.8",
    "google-genai>=1.68.0",
    "jupyter>=1.1.1",
    "pygame>=2.6.1",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
]
